service2/tournaments/serializers.py
from rest_framework import serializers
from .models import Tournament

class TournamentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tournament
        fields = ['creator_name', 'winner_nickname', 'participants_names']
